Add files via upload
The program calculates the time it takes for a particle dropped at some point on Earth´s surface to reach the exact point at the opposite part of Earth, using the Euler's method. It also plots the position of the particle, and the kinetic and potential energies associated to it as it goes through.